Move io_Write to Windos utils, where it belongs
