+++ examples/sality.py
@@ -11,6 +11,7 @@ sys.path.append("..")
 from qiling import Qiling
 from qiling.const import QL_VERBOSE
 from qiling.os.const import POINTER, DWORD, STRING, HANDLE
+from qiling.os.windows import utils
 from qiling.os.windows.api import *
 from qiling.os.windows.fncc import *
 from qiling.os.windows.dlls.kernel32.fileapi import _CreateFile
@@ -118,7 +119,7 @@ def hook_WriteFile(ql: Qiling, address: int, params):
     if hFile == 0x13371337:
         buffer = ql.mem.read(lpBuffer, nNumberOfBytesToWrite)
         try:
-            r, nNumberOfBytesToWrite = ql.amsint32_driver.os.io_Write(buffer)
+            r, nNumberOfBytesToWrite = utils.io_Write(ql.amsint32_driver, buffer)
             ql.mem.write(lpNumberOfBytesWritten, ql.pack32(nNumberOfBytesToWrite))
         except Exception as e:
             ql.log.exception("")
@@ -198,7 +199,7 @@ if __name__ == "__main__":
     ql.run(0x4053B2)
     ql.log.info("test kill thread")
     if ql.amsint32_driver:
-        ql.amsint32_driver.os.utils.io_Write(ql.pack32(0xdeadbeef))
+        utils.io_Write(ql.amsint32_driver, ql.pack32(0xdeadbeef))
 
         # TODO: Should stop at 0x10423, but for now just stop at 0x0001066a
         ql.amsint32_driver.hook_address(hook_stop_address, 0x0001066a)
