+++ qiling/os/utils.py
@@ -213,121 +213,6 @@ class QlOsUtils:
         # we want to rewrite the return address to the function
         self.ql.stack_write(0, start)
 
-    def io_Write(self, in_buffer):
-        heap = self.ql.os.heap
-
-        if self.ql.ostype == QL_OS.WINDOWS:
-
-            if self.ql.loader.driver_object.MajorFunction[IRP_MJ_WRITE] == 0:
-                # raise error?
-                return (False, None)
-
-        if self.ql.archbit == 32:
-            buf = self.ql.mem.read(self.ql.loader.driver_object.DeviceObject, ctypes.sizeof(DEVICE_OBJECT32))
-            device_object = DEVICE_OBJECT32.from_buffer(buf)
-        else:
-            buf = self.ql.mem.read(self.ql.loader.driver_object.DeviceObject, ctypes.sizeof(DEVICE_OBJECT64))
-            device_object = DEVICE_OBJECT64.from_buffer(buf)
-
-        alloc_addr = []
-        def build_mdl(buffer_size, data=None):
-            if self.ql.archtype == QL_ARCH.X8664:
-                mdl = MDL64()
-            else:
-                mdl = MDL32()
-
-            mapped_address = heap.alloc(buffer_size)
-            alloc_addr.append(mapped_address)
-            mdl.MappedSystemVa.value = mapped_address
-            mdl.StartVa.value = mapped_address
-            mdl.ByteOffset = 0
-            mdl.ByteCount = buffer_size
-            if data:
-                written = data if len(data) <= buffer_size else data[:buffer_size]
-                self.ql.mem.write(mapped_address, written)
-
-            return mdl
-        # allocate memory regions for IRP and IO_STACK_LOCATION
-        if self.ql.archtype == QL_ARCH.X8664:
-            irp_addr = heap.alloc(ctypes.sizeof(IRP64))
-            alloc_addr.append(irp_addr)
-            irpstack_addr = heap.alloc(ctypes.sizeof(IO_STACK_LOCATION64))
-            alloc_addr.append(irpstack_addr)
-            # setup irp stack parameters
-            irpstack = IO_STACK_LOCATION64()
-            # setup IRP structure
-            irp = IRP64()
-            irp.irpstack = ctypes.cast(irpstack_addr, ctypes.POINTER(IO_STACK_LOCATION64))
-        else:
-            irp_addr = heap.alloc(ctypes.sizeof(IRP32))
-            alloc_addr.append(irp_addr)
-            irpstack_addr = heap.alloc(ctypes.sizeof(IO_STACK_LOCATION32))
-            alloc_addr.append(irpstack_addr)
-            # setup irp stack parameters
-            irpstack = IO_STACK_LOCATION32()
-            # setup IRP structure
-            irp = IRP32()
-            irp.irpstack = ctypes.cast(irpstack_addr, ctypes.POINTER(IO_STACK_LOCATION32))
-
-        irpstack.MajorFunction = IRP_MJ_WRITE
-        irpstack.Parameters.Write.Length = len(in_buffer)
-        self.ql.mem.write(irpstack_addr, bytes(irpstack))
-
-        if device_object.Flags & DO_BUFFERED_IO:
-            # BUFFERED_IO
-            system_buffer_addr = heap.alloc(len(in_buffer))
-            alloc_addr.append(system_buffer_addr)
-            self.ql.mem.write(system_buffer_addr, bytes(in_buffer))
-            irp.AssociatedIrp.SystemBuffer.value = system_buffer_addr
-        elif device_object.Flags & DO_DIRECT_IO:
-            # DIRECT_IO
-            mdl = build_mdl(len(in_buffer))
-            if self.ql.archtype == QL_ARCH.X8664:
-                mdl_addr = heap.alloc(ctypes.sizeof(MDL64))
-            else:
-                mdl_addr = heap.alloc(ctypes.sizeof(MDL32))
-
-            alloc_addr.append(mdl_addr)
-
-            self.ql.mem.write(mdl_addr, bytes(mdl))
-            irp.MdlAddress.value = mdl_addr
-        else:
-            # NEITHER_IO
-            input_buffer_size = len(in_buffer)
-            input_buffer_addr = heap.alloc(input_buffer_size)
-            alloc_addr.append(input_buffer_addr)
-            self.ql.mem.write(input_buffer_addr, bytes(in_buffer))
-            irp.UserBuffer.value = input_buffer_addr
-
-        # everything is done! Write IRP to memory
-        self.ql.mem.write(irp_addr, bytes(irp))
-
-        # set function args
-        # TODO: make sure this is indeed STDCALL
-        self.ql.os.fcall = self.ql.os.fcall_select(STDCALL)
-        self.ql.os.fcall.writeParams(((POINTER, self.ql.loader.driver_object.DeviceObject), (POINTER, irp_addr)))
-
-        try:
-            # now emulate 
-            self.ql.run(self.ql.loader.driver_object.MajorFunction[IRP_MJ_WRITE])
-        except UcError as err:
-            verify_ret(self.ql, err)
-            
-        # read current IRP state
-        if self.ql.archtype == QL_ARCH.X8664:
-            irp_buffer = self.ql.mem.read(irp_addr, ctypes.sizeof(IRP64))
-            irp = IRP64.from_buffer(irp_buffer)
-        else:
-            irp_buffer = self.ql.mem.read(irp_addr, ctypes.sizeof(IRP32))
-            irp = IRP32.from_buffer(irp_buffer)
-
-        io_status = irp.IoStatus
-        # now free all alloc memory
-        for addr in alloc_addr:
-            # print("freeing heap memory at 0x%x" %addr) # FIXME: the output is not deterministic??
-            heap.free(addr)
-        return True, io_status.Information.value
-
     # Emulate DeviceIoControl() of Windows
     # BOOL DeviceIoControl(
     #      HANDLE       hDevice,
